# openapi.yaml
...
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5eb03bec82f82d390e783cd2
  title: Manage_Film_Example
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /film:
    post:
      x-skaffolder-id: 5eb03bee82f82d390e783d1b
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5eb03bed82f82d390e783cfd
      x-skaffolder-resource: Film
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Film
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Film"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                genre:
                  type: string
                title:
                  type: string
                year:
                  type: number
    get:
      x-skaffolder-id: 5eb03bee82f82d390e783d1f
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5eb03bed82f82d390e783cfd
      x-skaffolder-resource: Film
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Film
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Film"
  "/film/{id}":
    delete:
      x-skaffolder-id: 5eb03bee82f82d390e783d1e
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5eb03bed82f82d390e783cfd
      x-skaffolder-resource: Film
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Film
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5eb03bee82f82d390e783d18
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5eb03bed82f82d390e783cfd
      x-skaffolder-resource: Film
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Film
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Film"
    post:
      x-skaffolder-id: 5eb03bee82f82d390e783d1a
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5eb03bed82f82d390e783cfd
      x-skaffolder-resource: Film
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Film
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Film"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Film"
  "/film/findBycast/{key}":
    get:
      x-skaffolder-id: 5eb03bed82f82d390e783d05
      x-skaffolder-name: findBycast
      x-skaffolder-id-resource: 5eb03bed82f82d390e783cfd
      x-skaffolder-resource: Film
      x-skaffolder-crudAction: findBycast
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBycast
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBycast/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBycast
      tags:
        - Film
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Film"
  "/film/findByfilmMaker/{key}":
    get:
      x-skaffolder-id: 5eb03bee82f82d390e783d11
      x-skaffolder-name: findByfilmMaker
      x-skaffolder-id-resource: 5eb03bed82f82d390e783cfd
      x-skaffolder-resource: Film
      x-skaffolder-crudAction: findByfilmMaker
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findByfilmMaker
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findByfilmMaker/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findByfilmMaker
      tags:
        - Film
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Film"
  /filmmaker:
    post:
      x-skaffolder-id: 5eb03bee82f82d390e783d0c
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5eb03bed82f82d390e783cf1
      x-skaffolder-resource: FilmMaker
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - FilmMaker
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilmMaker"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                surname:
                  type: string
    get:
      x-skaffolder-id: 5eb03bee82f82d390e783d15
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5eb03bed82f82d390e783cf1
      x-skaffolder-resource: FilmMaker
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - FilmMaker
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FilmMaker"
  "/filmmaker/{id}":
    delete:
      x-skaffolder-id: 5eb03bee82f82d390e783d14
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5eb03bed82f82d390e783cf1
      x-skaffolder-resource: FilmMaker
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - FilmMaker
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5eb03bee82f82d390e783d0e
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5eb03bed82f82d390e783cf1
      x-skaffolder-resource: FilmMaker
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - FilmMaker
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilmMaker"
    post:
      x-skaffolder-id: 5eb03bee82f82d390e783d0d
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5eb03bed82f82d390e783cf1
      x-skaffolder-resource: FilmMaker
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - FilmMaker
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilmMaker"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FilmMaker"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Film:
      x-skaffolder-id: 5eb03bed82f82d390e783cfd
      x-skaffolder-id-db: 5eb03bec82f82d390e783cd3
      x-skaffolder-id-entity: 5eb03bed82f82d390e783cf3
      x-skaffolder-url: /film
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        genre:
          type: string
          x-skaffolder-id-attr: 5eb03bed82f82d390e783cf6
          x-skaffolder-type: String
          x-skaffolder-enumeration:
            - Action
            - Crime
            - Fantasy
            - Horror
        title:
          type: string
          x-skaffolder-id-attr: 5eb03bed82f82d390e783cf4
          x-skaffolder-type: String
          x-skaffolder-required: true
        year:
          type: number
          x-skaffolder-id-attr: 5eb03bed82f82d390e783cf5
          x-skaffolder-type: Number
      x-skaffolder-relations:
        cast:
          x-skaffolder-id: 5eb03bed82f82d390e783cf8
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 5eb03bed82f82d390e783cf3
          x-skaffolder-ent2: 5eb03bec82f82d390e783ce8
        filmMaker:
          x-skaffolder-id: 5eb03bed82f82d390e783cf7
          x-skaffolder-type: 1:m
          x-skaffolder-required: true
          x-skaffolder-ent1: 5eb03bed82f82d390e783cf3
          x-skaffolder-ent2: 5eb03bec82f82d390e783cee
      required:
        - _id
        - title
    FilmMaker:
      x-skaffolder-id: 5eb03bed82f82d390e783cf1
      x-skaffolder-id-db: 5eb03bec82f82d390e783cd3
      x-skaffolder-id-entity: 5eb03bec82f82d390e783cee
      x-skaffolder-url: /filmmaker
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        name:
          type: string
          x-skaffolder-id-attr: 5eb03bec82f82d390e783cef
          x-skaffolder-type: String
          x-skaffolder-required: true
        surname:
          type: string
          x-skaffolder-id-attr: 5eb03bec82f82d390e783cf0
          x-skaffolder-type: String
      x-skaffolder-relations: 
      required:
        - _id
        - name
...